+++ curves/src/bn256/fq12.rs
@@ -1,5 +1,5 @@
 use super::{fq::Fq, fq2::Fq2, fq6::Fq6};
-use crate::ff_ext::{
+use super::{
     quadratic::{QuadExtField, QuadExtFieldArith, QuadSparseMul},
     ExtField,
 };
